pyproject.toml
[tool.poetry]
name = "foodsto-api"
version = "0.1.0"
description = ""
authors = ["moh-ash96 <moh.ash96@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.8b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
